Raise if given precision is higher than 2 decimal places
sprintf '%0.2f' will use rounding when we format amounts for sending
to PayPal.

We guard inputs to the API to put the onus on the user to round or
truncate their own amounts so that there are no hidden surprises later.

This has no effect on users who are already sending amounts with 2
decimal places or less. Those who are relying on our library for
rounding (or earlier, truncating) will now receive ArgumentErrors.

While this usefully applies to currency amounts, all numeric inputs are
run through this check. However, I cannot see any other kind of value
that requires more precision.

Originally discussed in, and closes #19